- Refactor GitHub API routes into a plugin - Remove `orgs/github` member and repo API route files - Add `github` plugin with consolidated route handlers - Update `fresh.config.ts` to include `github` plugin - Adjust `TeamCards` component to use `github` plugin APIs
