+++ src/fresh.gen.ts
@@ -9,8 +9,6 @@
 
 import * as $_404 from "./routes/_404.tsx";
 import * as $_app from "./routes/_app.tsx";
-import * as $api_orgs_github_members from "./routes/api/orgs/github/members.ts";
-import * as $api_orgs_github_repos from "./routes/api/orgs/github/repos.ts";
 import * as $dashboard from "./routes/dashboard.tsx";
 import * as $index from "./routes/index.tsx";
 import * as $login from "./routes/login.tsx";
@@ -26,8 +24,6 @@ const manifest = {
   routes: {
     "./routes/_404.tsx": $_404,
     "./routes/_app.tsx": $_app,
-    "./routes/api/orgs/github/members.ts": $api_orgs_github_members,
-    "./routes/api/orgs/github/repos.ts": $api_orgs_github_repos,
     "./routes/dashboard.tsx": $dashboard,
     "./routes/index.tsx": $index,
     "./routes/login.tsx": $login,
