+++ src/plugins/github.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2025. Xodium.
+ * All rights reserved.
+ */
+
+import {Plugin} from "$fresh/server.ts";
+import {createOrgDataHandler} from "../utils/utils.ts";
+
+export interface Member {
+  login: string;
+  avatar_url: string;
+  html_url: string;
+}
+
+export interface Repo {
+  id: number;
+  name: string;
+  full_name: string;
+  html_url: string;
+  description: string;
+}
+
+export default {
+  name: "github",
+  routes: [
+    {
+      path: "/api/github/org/members",
+      handler: createOrgDataHandler<Member[]>(
+        "members",
+        "/orgs/{org}/members",
+      ),
+    },
+    {
+      path: "/api/github/org/repos",
+      handler: createOrgDataHandler<Repo[]>(
+        "repos",
+        "/orgs/{org}/repos",
+      ),
+    },
+  ],
+} as Plugin;
